FlaskRunner/JexxlAPI/requirements.txt
langchain-community
langchain-openai
openai
flask
waitress
python-dotenv
requests
python-Levenshtein
fuzzywuzzy